refactor(error): create module for the AstarteError
This reduces the code complexity of the main lib file
