+++ README.md
@@ -29,8 +29,8 @@ for more information regarding Astarte and the available SDKs.
 use astarte_device_sdk::{
     database::AstarteSqliteDatabase,
     options::AstarteOptions,
-    AstarteDeviceSdk,
-    AstarteError
+    error::AstarteError,
+    AstarteDeviceSdk
 };
 
 async fn run_astarte_device() -> Result<(), AstarteError> {
